Merge data sets
The goal of this code is to upload and reconcile primary differences in data structure and nomenclature from datasets from two different sources.  
1. Maternal mortality: https://ghdx.healthdata.org/record/ihme-data/maternal-mortality-estimates-and-mdg-5-attainment-country-1990-2011 
2. Literacy rates for women: https://ourworldindata.org/grapher/adult-literacy-female. 

I rename columns and write functions to rename "global" to faciliate merging. 

Once that's complete, the datasets are merged by country and year and I write it into an excel output for next steps with modeling and Shiny development.