Natalie's code for final markdown pdf
This code pulls together all the data and writes out the introduction and context for the subsequent plots. It prints to a PDF with all the plots. 
I have the introduction with links to this github and pull requests, and topic specific context on this analysis. 

I include the merging data code and data prep pieces. I have 4 plots at a global and regional level across MMR and female literacy over time (1990-2011)
1. Global MMR over time
2. Global female literacy rate over time
3. Global MMR vs. female literacy by time

From these plots we dig into the regional trend: 
4. GBD regional trend of MMR vs. female literacy rates

Below each plot I list key takeaways, which can be summarized with everyone's visualizations in a conclusion section.